+++ comp2013/src/main/java/comp2013/View/SnakeScenes/GameSelectScene.java
@@ -1,4 +1,0 @@
-package comp2013.View.SnakeScenes;
-
-public class GameSelectScene {
-}